+++ packages/babel-preset-gatsby/src/index.js
@@ -2,9 +2,11 @@ const path = require(`path`)
 
 const resolve = m => require.resolve(m)
 
+const IS_TEST = (process.env.BABEL_ENV || process.env.NODE_ENV) === `test`
+
 const loadCachedConfig = () => {
   let pluginBabelConfig = {}
-  if (process.env.NODE_ENV !== `test`) {
+  if (!IS_TEST) {
     try {
       pluginBabelConfig = require(path.join(
         process.cwd(),
@@ -97,7 +99,20 @@ module.exports = function preset(_, options = {}) {
           absoluteRuntimePath,
         },
       ],
-      ...plugins,
-    ],
+      [
+        resolve(`@babel/plugin-transform-spread`),
+        {
+          loose: false, // Fixes #14848
+        },
+      ],
+      IS_TEST && resolve(`babel-plugin-dynamic-import-node`),
+      stage === `build-javascript` && [
+        // Remove PropTypes from production build
+        resolve(`babel-plugin-transform-react-remove-prop-types`),
+        {
+          removeImport: true,
+        },
+      ],
+    ].filter(Boolean),
   }
 }