+++ packages/gatsby-plugin-manifest/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-manifest@2.2.2...gatsby-plugin-manifest@2.2.3) (2019-07-12)
+
+### Bug Fixes
+
+- correct links in package changelogs ([#15630](https://github.com/gatsbyjs/gatsby/issues/15630)) ([d07b9dd](https://github.com/gatsbyjs/gatsby/commit/d07b9dd))
+
 ## [2.2.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-manifest@2.2.1...gatsby-plugin-manifest@2.2.2) (2019-07-11)
 
 **Note:** Version bump only for package gatsby-plugin-manifest