+++ packages/gatsby-plugin-mdx/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.11...gatsby-plugin-mdx@1.0.12) (2019-07-12)
+
+### Bug Fixes
+
+- **gatsby-plugin-mdx:** Fix Safari 9 `const` syntax error ([#15644](https://github.com/gatsbyjs/gatsby/issues/15644)) ([a823463](https://github.com/gatsbyjs/gatsby/commit/a823463))
+
 ## [1.0.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.10...gatsby-plugin-mdx@1.0.11) (2019-07-11)
 
 ### Bug Fixes