+++ packages/gatsby-source-npm-package-search/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@2.1.1...gatsby-source-npm-package-search@2.1.2) (2019-07-12)
+
+**Note:** Version bump only for package gatsby-source-npm-package-search
+
 ## [2.1.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@2.1.0...gatsby-source-npm-package-search@2.1.1) (2019-07-11)
 
 **Note:** Version bump only for package gatsby-source-npm-package-search