+++ packages/gatsby-transformer-remark/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-remark@2.6.3...gatsby-transformer-remark@2.6.4) (2019-07-12)
+
+### Bug Fixes
+
+- correct links in package changelogs ([#15630](https://github.com/gatsbyjs/gatsby/issues/15630)) ([d07b9dd](https://github.com/gatsbyjs/gatsby/commit/d07b9dd))
+
 ## [2.6.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-remark@2.6.2...gatsby-transformer-remark@2.6.3) (2019-07-11)
 
 **Note:** Version bump only for package gatsby-transformer-remark